consertando os erros, todos ids devem ser long

+++ src/main/java/ufrn/br/repository/ProdutoCategoriaRepository.java
@@ -3,5 +3,5 @@ package ufrn.br.repository;
 import ufrn.br.model.ProdutoCategoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdutoCategoriaRepository extends JpaRepository<ProdutoCategoria, Integer> {
+public interface ProdutoCategoriaRepository extends JpaRepository<ProdutoCategoria, Long> {
 }
